Moved database.py to init.py.
